Move OpenAPI extensions to routing assembly
