+++ src/Http/Headers/src/MediaTypeCollection.cs
