+++ src/Mvc/Mvc.Core/src/ApiExplorer/IApiResponseMetadataProvider.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using Microsoft.AspNetCore.Http.Metadata;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.Mvc.Formatters;
 
@@ -11,21 +12,21 @@ namespace Microsoft.AspNetCore.Mvc.ApiExplorer
     /// Provides a return type, status code and a set of possible content types returned by a
     /// successful execution of the action.
     /// </summary>
-    public interface IApiResponseMetadataProvider : IFilterMetadata
+    public interface IApiResponseMetadataProvider : IFilterMetadata, IProducesResponseTypeMetadata
     {
         /// <summary>
         /// Gets the optimistic return type of the action.
         /// </summary>
-        Type? Type { get; }
+        new Type? Type { get; }
 
         /// <summary>
         /// Gets the HTTP status code of the response.
         /// </summary>
-        int StatusCode { get; }
+        new int StatusCode { get; }
 
         /// <summary>
         /// Configures a collection of allowed content types which can be produced by the action.
         /// </summary>
-        void SetContentTypes(MediaTypeCollection contentTypes);
+        new void SetContentTypes(MediaTypeCollection contentTypes);
     }
 }
